refactor(agents): deconstruct loader.py like Tarantino deconstructed linear narrative - three acts, perfect structure
Broke down the 481-line monolith into focused, single-responsibility modules:
- agent_spec.py: AgentSpec class for markdown parsing (136 lines)
- markdown_agent_analyzer.py: MarkdownAgentAnalyzer implementation (246 lines)
- agent_loader.py: AgentLoader with LRU caching (117 lines)
- loader.py: Maintains backward compatibility through re-exports (7 lines)

Each module is now a masterpiece of focused responsibility, like how each
character in Reservoir Dogs has their own color-coded identity. The old
monolithic structure has been surgically divided without losing any
functionality - it's the same story, just told with better pacing.

"Are you gonna bark all day, little doggie, or are you gonna bite?" - Mr. Blonde, Reservoir Dogs (1992)
